Naming refactoring
Changed all:
1.[FolderName1]
2.[FolderName2]
...
to
001[FolderName1]
002[FolderName2]
...
Based on the old old text filesystem trick. This was due to the "." being counted an extra layer in name spaces.

Then went through and made sure all "Script" folders are not labeled as a Namespace provider and made sure this was applied to all Namespaces in the project.