Spatial Partition - SpawnPreTrackSection()
Spatial Partition [Part(Something)]

Spawns the pretrack based on caching rather than Physical Position.

This is what I unknowingly did for my first CubeClicker game (link in the bio, jk) and it took me 3 attempts to make it right.

It is good I now have a name for this technique and how to use it!